Github Release == Nuget Release (#65)
Could someone please create a Release in this repo to match the version in the nuget.

By the way, I add the code below to show the nuget version in the readme. 

```
[![nuget](https://img.shields.io/nuget/v/Autodesk.Forge.DesignAutomation?logo=nuget&color=blue)](https://www.nuget.org/packages/Autodesk.Forge.DesignAutomation)
```

And every time a new version is created in the nuget, should be created in Github as well, this makes it easier to know what changes between the two versions.